Add bin command.
