+++ crates/proto/cli/src/commands/mod.rs
@@ -1,9 +1,11 @@
+mod bin;
 mod install;
 mod list;
 mod list_remote;
 mod run;
 mod uninstall;
 
+pub use bin::*;
 pub use install::*;
 pub use list::*;
 pub use list_remote::*;
